Password feature in Users module
Fixes #1736
