+++ src/OrchardCore.Modules/OrchardCore.Users/Controllers/AccountController.cs
@@ -34,34 +34,20 @@ namespace OrchardCore.Users.Controllers
         private readonly UserManager<IUser> _userManager;
         private readonly ILogger _logger;
         private readonly ISiteService _siteService;
-        private readonly ISmtpService _smtpService;
-        private readonly IShapeFactory _shapeFactory;
-        private readonly IHtmlDisplay _displayManager;
 
         public AccountController(
             IUserService userService,
             SignInManager<IUser> signInManager,
             UserManager<IUser> userManager,
             ILogger<AccountController> logger,
-            ISiteService siteService,
-            ISmtpService smtpService,
-            IShapeFactory shapeFactory,
-            IHtmlDisplay displayManager,
-            IStringLocalizer<AccountController> stringLocalizer)
+            ISiteService siteService)
         {
             _signInManager = signInManager;
             _userManager = userManager;
             _userService = userService;
             _logger = logger;
             _siteService = siteService;
-            _smtpService = smtpService;
-            _shapeFactory = shapeFactory;
-            _displayManager = displayManager;
-
-            T = stringLocalizer;
         }
-
-        IStringLocalizer T { get; set; }
 
         [HttpGet]
         [AllowAnonymous]
@@ -212,112 +198,12 @@ namespace OrchardCore.Users.Controllers
         [AllowAnonymous]
         public async Task<IActionResult> ForgotPassword()
         {
-            if (!(await _siteService.GetSiteSettingsAsync()).As<RegistrationSettings>().EnableLostPassword)
+            if (!(await _siteService.GetSiteSettingsAsync()).As<PasswordSettings>().EnableLostPassword)
             {
                 return NotFound();
             }
 
             return View();
-        }
-
-        [HttpPost]
-        [AllowAnonymous]
-        public async Task<IActionResult> ForgotPassword(ForgotPasswordViewModel model)
-        {
-            if (!(await _siteService.GetSiteSettingsAsync()).As<RegistrationSettings>().EnableLostPassword)
-            {
-                return NotFound();
-            }
-
-            if (ModelState.IsValid)
-            {
-                var user = (User)await _userService.GetForgotPasswordUserAsync(model.UserIdentifier);
-                if (user != null)
-                {
-                    user.ResetToken = Convert.ToBase64String(Encoding.UTF8.GetBytes(user.ResetToken));
-                    // send email with callback link
-                    await SendResetTokenAsync(user);
-                }
-            }
-
-            // returns to confirmation page anyway: we don't want to let scrapers know if a username or an email exist
-            return RedirectToLocal(Url.Action("ForgotPasswordConfirmation"));
-        }
-
-        [HttpGet]
-        [AllowAnonymous]
-        public IActionResult ForgotPasswordConfirmation()
-        {
-            return View();
-        }
-
-        [HttpGet]
-        [AllowAnonymous]
-        public async Task<IActionResult> ResetPassword(string code = null)
-        {
-            if (!(await _siteService.GetSiteSettingsAsync()).As<RegistrationSettings>().EnableLostPassword)
-            {
-                return NotFound();
-            }
-            if (code == null)
-            {
-                //"A code must be supplied for password reset.";
-            }
-            return View(new ResetPasswordViewModel { ResetToken = code });
-        }
-
-        [HttpPost]
-        [AllowAnonymous]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> ResetPassword(ResetPasswordViewModel model)
-        {
-            if (!(await _siteService.GetSiteSettingsAsync()).As<RegistrationSettings>().EnableLostPassword)
-            {
-                return NotFound();
-            }
-
-            if (ModelState.IsValid)
-            {
-                if (await _userService.ResetPasswordAsync(model.Email, Encoding.UTF8.GetString(Convert.FromBase64String(model.ResetToken)), model.NewPassword, (key, message) => ModelState.AddModelError(key, message)))
-                {
-                    return RedirectToLocal(Url.Action("ResetPasswordConfirmation"));
-                }
-            }
-
-            return View(model);
-        }
-
-        [HttpGet]
-        [AllowAnonymous]
-        public IActionResult ResetPasswordConfirmation()
-        {
-            return View();
-        }
-
-        protected async Task<bool> SendResetTokenAsync(User user)
-        {
-            var viewName = "TemplateUserLostPassword";
-            var model = Arguments.From(new { User = user, LostPasswordUrl = Url.Action("ResetPassword", "Account", new { code = user.ResetToken }, HttpContext.Request.Scheme) });
-
-            var options = ControllerContext.HttpContext.RequestServices.GetRequiredService<IOptions<MvcViewOptions>>();
-            ControllerContext.RouteData.Values["action"] = viewName;
-            ControllerContext.RouteData.Values["controller"] = "";
-            var viewEngineResult = options.Value.ViewEngines.Select(x => x.FindView(ControllerContext, viewName, true)).FirstOrDefault(x => x != null);
-            var displayContext = new DisplayContext()
-            {
-                ServiceProvider = ControllerContext.HttpContext.RequestServices,
-                Value = await _shapeFactory.CreateAsync(viewName, model),
-                ViewContext = new ViewContext(ControllerContext, viewEngineResult.View, ViewData, TempData, new StringWriter(), new HtmlHelperOptions())
-            };
-            var htmlContent = await _displayManager.ExecuteAsync(displayContext);
-
-            var message = new MailMessage() { Body = WebUtility.HtmlDecode(htmlContent.ToString()), IsBodyHtml = true, Subject = T["Reset your password"] };
-            message.To.Add(user.Email);
-
-            // send email
-            var result = await _smtpService.SendAsync(message);
-
-            return result.Succeeded;
         }
     }
 }
