+++ src/OrchardCore.Modules/OrchardCore.Users/Models/PasswordSettings.cs
@@ -1,0 +1,7 @@
+namespace OrchardCore.Users.Models
+{
+    public class PasswordSettings
+    {
+        public bool EnableLostPassword { get; set; }
+    }
+}
