+++ src/OrchardCore.Modules/OrchardCore.Users/Models/RegistrationSettings.cs
@@ -3,6 +3,5 @@ namespace OrchardCore.Users.Models
     public class RegistrationSettings
     {
         public bool UsersCanRegister { get; set; }
-        public bool EnableLostPassword { get; set; }
     }
 }
