+++ src/OrchardCore.Modules/OrchardCore.Users/Views/PasswordSettings.Edit.cshtml
@@ -1,0 +1,6 @@
+@model OrchardCore.Users.Models.PasswordSettings
+
+<fieldset class="form-group" asp-validation-class-for="EnableLostPassword">
+    <input asp-for="EnableLostPassword" type="checkbox" />
+    <label asp-for="EnableLostPassword">@T["Display a link to enable users to reset their password"] <span asp-validation-for="EnableLostPassword"></span></label>
+</fieldset>
